backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.40
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pytest==8.3.5
httpx==0.28.1
